Update UserController with Tests and UserService with Tests

+++ backend/src/main/java/org/example/backend/controller/UserController.java
@@ -40,6 +40,9 @@ public class UserController {
     @PutMapping("/{userId}/updateFavoriteProjectsOfUser/{projectId}")
     public User updateUserById(@PathVariable String userId, @PathVariable String projectId) {
         User user = service.getUserById(userId);
+        if (user == null) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "User not found");
+        }
         if(user.getFavoriteProjects().contains(projectId)) {
             return service.removeFavoriteProject(user, projectId);
         } else {
